Complete permissions mutation

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -45,9 +45,17 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
     // 2. check if they own that item, or have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permission.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
+
     // 3. delete
     return ctx.db.mutation.deleteItem({ where }, info);
   },
